fix: move otel setup to main.ts
